# tyst-api-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tyst_api_rest",
    "description": "TYST REST API",
    "license": {
      "name": "Apache-2.0 WITH FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/mac/{algorithm}/keygen": {
      "post": {
        "tags": [
          "mac_resources"
        ],
        "summary": "Generation of MAC key.",
        "description": "Generate a suitable key for the Message Authentication Code (MAC) algorithm.",
        "operationId": "mac_keygen",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "description": "The MAC algorithm.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "key"
                  ],
                  "properties": {
                    "key": {
                      "type": "object",
                      "required": [
                        "key_material_b64"
                      ],
                      "properties": {
                        "confinement_id": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Identifier used by provider to locate the correct confinement."
                        },
                        "confinement_type": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Type of confinement. E.g. \"soft\"."
                        },
                        "key_material_b64": {
                          "type": "array",
                          "items": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 0
                          },
                          "description": "Base64 encoded key material"
                        },
                        "object_reference": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Identifier used by provider to locate the correct object."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Key generation failed or was never started."
          }
        }
      }
    },
    "/mac/{algorithm}/mac": {
      "post": {
        "tags": [
          "mac_resources"
        ],
        "summary": "Produce MAC",
        "description": "Produce Message Authentication Code (MAC).",
        "operationId": "mac",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "description": "The MAC algorithm.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "key",
                  "message_b64"
                ],
                "properties": {
                  "key": {
                    "type": "object",
                    "required": [
                      "key_material_b64"
                    ],
                    "properties": {
                      "confinement_id": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Identifier used by provider to locate the correct confinement."
                      },
                      "confinement_type": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Type of confinement. E.g. \"soft\"."
                      },
                      "key_material_b64": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "format": "int32",
                          "minimum": 0
                        },
                        "description": "Base64 encoded key material"
                      },
                      "object_reference": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Identifier used by provider to locate the correct object."
                      }
                    }
                  },
                  "message_b64": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "mac_b64"
                  ],
                  "properties": {
                    "mac_b64": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Fail"
          }
        }
      }
    },
...
    "/prng/{algorithm}/random/{length}": {
      "get": {
        "tags": [
          "prng_resources"
        ],
        "summary": "Get random bytes.",
        "description": "Return random bytes of the requested length using and instance of the\nrequested Psudedo Random Number Generator (PRNG) algorithm.\nIf 'algorithm_name' is set to 'default', a cryptographically strong\ngenerator will be choosen.\nSending 'Accept: */*' returns the hex encoded version of the digest.",
        "operationId": "prng_random",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "length",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Random data",
            "content": {
              "application/octet-stream": {}
            }
          },
          "400": {
            "description": "Failure"
          }
        }
      }
    },
...
  },
...
}